Vertical Navbar initial construction
Constructed vertical navbar similarly to “Navbar.js” (which represents the horizontal navbar). Added a component "VNavbar" to components folder with similar "vMenuitems.js" and "vNavbar.css" files. Added relevant names of links to “vMenuitems”.js and toyed with “VNavbar.css” until it was in a decent place.
Also added a <div> tag for a corner logo inserted in the top left of the canvas of "Home.js" not sure if needed, but we can always revert. Would've worked better if I could figure out how to resize it well.