Backend stores new hatchery details from home.js popup
The backend now stores new Hatchery details with the next button in the pop-up window. I went ahead and calculated the volume from the dimensions so we could have it in the database as a number rather than a string. The function deletes all non-numeric characters and uses logic to figure out which numbers should be multiplied together, that being said it could break if Akissi changes the dimensions out, I tried to comment how it works in Modal.js here’s an example ("17” L x 6” W x 10” H" ---> 17610 ---> 17 * 6 * 10 ---> 1020). I did something similar to enter the number of grubs as a number as well. I also updated the hatchery model to reflect the changed variable types.